add dir database for database files and move others files in backend to lib except main.py
