+++ backend/lib/datasources/providers/nutrition_fake.py
@@ -1,4 +1,4 @@
-from lib.service.interfaces import nutrition 
+from service.interfaces import nutrition 
 
 
 class FakeNutritionProvider(nutrition.NutritionProvider):
